Move crypto implementations to src/crypto/
